completed 4.21

+++ part4/blogList-backend/controllers/blogs.js
@@ -1,5 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/users')
 
 
 blogsRouter.get('/', async(request, response) => {
@@ -11,8 +12,7 @@ blogsRouter.post('/', async(request, response,next) => {
   const body = request.body
   const token = request.token
   const user = request.user
-  console.log("user", user)
-  console.log("token", token)
+  
   if (!(token && user)) {
     return response.status(401).json({ error: 'token missing or invalid' }) 
   }
@@ -33,7 +33,6 @@ blogsRouter.post('/', async(request, response,next) => {
   } else {
     const savedBlog = await blog.save()
     
-    console.log("blogId", savedBlog)
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
 
@@ -42,10 +41,24 @@ blogsRouter.post('/', async(request, response,next) => {
    
 })
 
+
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
-  await Blog.findByIdAndDelete(id)
-  response.status(204).end()
+  const token = request.token
+  const user = request.user
+  const blog = await Blog.findById(id)
+  if (!(token && user)) {
+    return response.status(401).json({error:"token missing or invalid"})
+  }
+  if (!blog) {
+    return response.status(400).json({error:"couldn't find blog"})
+  }
+  if (blog.user.toString() === user.id.toString()) {
+    await User.findByIdAndUpdate(user.id, { $pull: { blogs: id } }) //using pull method from mongoose to delete(remove) the blog which need to remved from database
+    await Blog.findByIdAndDelete(id)
+    
+    response.status(204).json(blog).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
